Remove attach_file and mime_types dependency
Be ruthless on dependencies. We don’t need this. While the method is
nice to have it hasn’t proved any use for me. We can bring it back if
necessary but for now avoiding another dependency. We can have this
method ask for file path and mime type. Updated sample code.